requirements.txt
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
h5py>=3.8.0
requests>=2.28.0
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
gevent>=22.10.0
gevent-websocket>=0.10.1
joblib>=1.3.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
sseclient-py>=1.8.0
